COM-1702 create activitiesList component
