+++ src/components/ActivitiesList/styles.js
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native';
+import { MARGIN, PADDING } from '../../styles/metrics';
+
+const styles = StyleSheet.create({
+  activityCellList: {
+    marginTop: MARGIN.MD,
+    paddingHorizontal: PADDING.MD,
+  },
+  separator: {
+    marginHorizontal: MARGIN.XS,
+  },
+});
+
+export default styles;
